requirements_min.txt
Flask==2.2.2
Nuitka==1.3.7
ordered-set==4.1.0
pythonnet==3.0.1
pywebview==3.7.2
zstandard==0.19.0